chore: refactor statement parsers to re-use more code
Combine some of the methods in the statement parsers into one method in
the abstract base class, so more code can be re-used. This will also be
used to support keeping the comments that are sent to the JDBC driver,
so the comments can also be sent to Spanner.
